+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Display orchestrator and chaincode versions in node info command
 - Accept full datamanager, metrics and parent_tasks objects in get_*_tuple responses
 - Add task extra information related to start_date and end_date
+- Cyclic strategy example
 
 ### Changed